Feature/add context to thread worker (#7)
* Update create thread and run, add list run endpoint

* Handle exceptions correctly in the endpoint response

* Remove get run by id from context due to skip-locked behavior

* Add context to thread worker

* Update async worker to handle sychronous run and make run skip-locked

* Update run last error format

---------

+++ llamphouse/core/routes/message.py
@@ -9,6 +9,10 @@ router = APIRouter()
 @router.post("/threads/{thread_id}/messages", response_model=MessageObject)
 async def create_message(thread_id: str, request: CreateMessageRequest):
     try:
+        thread = db.get_thread_by_id(thread_id)
+        if not thread:
+            raise HTTPException(status_code=404, detail="Thread not found.")
+        
         message = db.insert_message(thread_id, request)
         
         return MessageObject(
@@ -31,12 +35,18 @@ async def create_message(thread_id: str, request: CreateMessageRequest):
                     created_at=int(message.created_at.timestamp()),
                     thread_id=message.thread_id
                 ) 
+    except HTTPException as http_exc:
+        raise http_exc
     except Exception as e:
-        raise HTTPException(status_code=500, detail=str(e))
+        raise HTTPException(status_code=500, detail=f"Internal Server Error: {str(e)}")
 
 @router.get("/threads/{thread_id}/messages", response_model=MessageListResponse)
 async def list_messages(thread_id: str, limit: int = 20, order: str = "desc", after: Optional[str] = None, before: Optional[str] = None):
     try:
+        thread = db.get_thread_by_id(thread_id)
+        if not thread:
+            raise HTTPException(status_code=404, detail="Thread not found.")
+        
         messages = db.get_messages_by_thread_id(
             thread_id=thread_id,
             limit=limit + 1,
@@ -77,12 +87,18 @@ async def list_messages(thread_id: str, limit: int = 20, order: str = "desc", af
             last_id=last_id,
             has_more=has_more
         )
+    except HTTPException as http_exc:
+        raise http_exc
     except Exception as e:
-        raise HTTPException(status_code=500, detail=str(e))
+        raise HTTPException(status_code=500, detail=f"Internal Server Error: {str(e)}")
 
 @router.get("/threads/{thread_id}/messages/{message_id}", response_model=MessageObject)
 async def retrieve_message(thread_id: str, message_id: str):
     try:
+        thread = db.get_thread_by_id(thread_id)
+        if not thread:
+            raise HTTPException(status_code=404, detail="Thread not found.")
+        
         message = db.get_message_by_id(message_id)
         if not message:
             raise HTTPException(status_code=404, detail="Message not found.")
@@ -107,8 +123,10 @@ async def retrieve_message(thread_id: str, message_id: str):
                     created_at=int(message.created_at.timestamp()),
                     thread_id=message.thread_id
                 ) 
+    except HTTPException as http_exc:
+        raise http_exc
     except Exception as e:
-        raise HTTPException(status_code=500, detail=str(e))
+        raise HTTPException(status_code=500, detail=f"Internal Server Error: {str(e)}")
 
 @router.post("/threads/{thread_id}/messages/{message_id}", response_model=MessageObject)
 async def modify_message(thread_id: str, message_id: str, request: ModifyMessageRequest):
@@ -137,12 +155,22 @@ async def modify_message(thread_id: str, message_id: str, request: ModifyMessage
                     created_at=int(message.created_at.timestamp()),
                     thread_id=message.thread_id
                 ) 
+    except HTTPException as http_exc:
+        raise http_exc
     except Exception as e:
-        raise HTTPException(status_code=500, detail=str(e))
+        raise HTTPException(status_code=500, detail=f"Internal Server Error: {str(e)}")
 
 @router.delete("/threads/{thread_id}/messages/{message_id}", response_model=DeleteMessageResponse)
 async def delete_message(thread_id: str, message_id: str):
     try:
+        thread = db.get_thread_by_id(thread_id)
+        if not thread:
+            raise HTTPException(status_code=404, detail="Thread not found.")
+        
+        message = db.get_message_by_id(message_id)
+        if not message:
+            raise HTTPException(status_code=404, detail="Message not found.")
+
         deleted = db.delete_message(thread_id, message_id)
         if not deleted:
             return DeleteMessageResponse(
@@ -154,5 +182,7 @@ async def delete_message(thread_id: str, message_id: str):
             id=thread_id,
             deleted=True
         )
+    except HTTPException as http_exc:
+        raise http_exc
     except Exception as e:
-        raise HTTPException(status_code=500, detail=str(e))
+        raise HTTPException(status_code=500, detail=f"Internal Server Error: {str(e)}")
